requirements.txt
flask
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage==2.10.0
google-auth
pandas
numpy==1.24.3
matplotlib
seaborn
statsmodels
scikit-learn==1.3.0
db-dtypes